Move refactor admin pages in admin directory
